PlaybackLink (renamed) now just takes a URL instead of the entire album
